requirements.txt
feedparser
python-telegram-bot==13.15